python/requirements.txt
spotipy>=2.23.0
Pillow>=10.0.0
requests>=2.31.0
pyserial>=3.5
python-dotenv>=1.0.0
numpy>=1.24.0
Wand>=0.6.11
